Extract classes to separate classes
